Remove deprecated ErrorMessage::Translator mechanism and refactor
Rely solely on the new mechanism, remove the old code
and apply further refactorings that then become possible.

This includes moving the helpers to the sole classes that now
use them and deleting the helper
